Get sysio-writeread example working again
This commit includes a wide range of fixes and code cleanup.
Highlights follow:
- disable incorrect EOF calculation caused by local metadata
  having incorrect total file size
- replace int/long with size_t where appropriate
- fix qsort comparison functions to work for unsigned types
- make sure clients and server use same data structs
  for shared memory exchanges
- make sure reqmgr and svcmgr use same data structs
  for MPI exchanges
- zero static buffers before reuse
- improve key/value access macros in MDHIM code
- updated constants to reflect reasonable buffer sizes
- add PMIx configuration option for build even if available
- more comments to improve code understanding

Checkpatch Skip Notes:
- LOG macro can't meet checkpatch criteria without breaking it
- not applying UnifyCR style to meta/src files

TEST_CHECKPATCH_SKIP_FILES=server/src/log.h
TEST_CHECKPATCH_SKIP_FILES+=,meta/src/ds_leveldb.c
TEST_CHECKPATCH_SKIP_FILES+=,meta/src/ds_leveldb.h
TEST_CHECKPATCH_SKIP_FILES+=,meta/src/range_server.c
TEST_CHECKPATCH_SKIP_FILES+=,meta/src/range_server.h

+++ client/src/unifycr-sysio.h
@@ -108,7 +108,7 @@ int unifycr_fd_read(int fd, off_t pos, void *buf, size_t count,
  * allocates new bytes and updates file size as necessary,
  * fills any gaps with zeros */
 int unifycr_fd_write(int fd, off_t pos, const void *buf, size_t count);
-int unifycr_fd_logreadlist(read_req_t *read_req, int count);
+int unifycr_fd_logreadlist(read_req_t *read_req, size_t count);
 int compare_read_req(const void *a, const void *b);
 int compare_index_entry(const void *a, const void *b);
